tests/requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
requests>=2.31.0
boto3>=1.28.0
moto>=4.2.0
python-jose[cryptography]>=3.3.0
Pillow>=10.0.0
jsonschema>=4.17.0
